+++ src/frontend/servers/y-provider/src/routes.ts
@@ -2,5 +2,6 @@ export const routes = {
   COLLABORATION_WS: '/collaboration/ws/',
   COLLABORATION_RESET_CONNECTIONS: '/collaboration/api/reset-connections/',
   CONVERT: '/api/convert/',
+  CONTENT: '/api/content/',
   COLLABORATION_GET_CONNECTIONS: '/collaboration/api/get-connections/',
 };